tentative gestion de signal
