+++ includes/signals.h
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   signals.h                                          :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: yseguin <yseguin@student.42perpignan.fr    +#+  +:+       +#+        */
+/*   By: yseguin <youvataque@icloud.com>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/02/19 18:04:15 by ilbonnev          #+#    #+#             */
-/*   Updated: 2025/03/05 15:47:54 by yseguin          ###   ########.fr       */
+/*   Updated: 2025/03/27 14:39:29 by yseguin          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,4 +15,7 @@
 
 typedef struct s_shell	t_shell;
 
+/* handler for signals */
+void	handle_sigint(int sig);
+void	handle_sigquit(int sig);
 #endif
