+++ src/signals/handler.c
@@ -1,0 +1,31 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   handler.c                                          :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: yseguin <youvataque@icloud.com>            +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2025/03/27 14:38:25 by yseguin           #+#    #+#             */
+/*   Updated: 2025/03/27 14:38:56 by yseguin          ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include "../includes/minishell.h"
+
+///////////////////////////////////////////////////////////////////////////////
+// rewrite shell for ctrl c
+void	handle_sigint(int sig)
+{
+	(void)sig;
+	write(STDOUT_FILENO, "\n", 1);
+	rl_on_new_line();
+	rl_replace_line("", 0);
+	rl_redisplay();
+}
+
+///////////////////////////////////////////////////////////////////////////////
+// rewrite shell for ctrl c
+void	handle_sigquit(int sig)
+{
+	(void)sig;
+}
